tweak colors
Made the colors a little more vibrant. Game background is now dark. Feels like it's easier on the eyes if one is staring at the screen for a long time.

Made the "power" input a "reset" button.
Set mode to strict by default and got rid of the strict toggle in an attempt to simplify things.